Parallelised, tidied up and documented the new demosaicing implementation. Integrated the new bayer decode based on the availability of std::par functionality. On older compilers it will revert to the old bayer algorithm. Also parallelised the monochrome algorithm for situations where it is not a simple memcpy but requires type conversion and bit shifting.

+++ include/psinc/handlers/DataHandler.hpp
@@ -16,7 +16,7 @@ namespace psinc
 			virtual ~DataHandler() {}
 
 			/// Process the data in the supplied buffer using the known width and height of the image
-			virtual bool Process(bool monochrome, const bool hdr, const std::vector<emg::byte> &data, const int width, const int height, const emg::byte bayerMode) = 0;
+			virtual bool Process(bool monochrome, const bool hdr, const std::vector<emg::byte> &data, const size_t width, const size_t height, const emg::byte bayerMode) = 0;
 
 			// This should only be written to by the acquisition system, but can be read
 			// from outside to know when the USB transport has prepped the camera for
